requirements.txt
pytesseract
opencv-python
python-dotenv
